Upgrade eslint-config-prettier from ^2 to ^6
Changelog: https://github.com/prettier/eslint-config-prettier/blob/master/CHANGELOG.md

This eslint config *turns off* eslint rules that are formatting-specific
and might conflict with prettier's formatting. The idea is that if you are using
eslint-plugin-prettier, you want prettier *alone* (not eslint also) to be
responsible for formatting and the linting of formatting.

There are a handful of tweaks and exceptions to rules that are added in the
changelog. Despite the large-ish version jump here there are not that
many changes that look to me to be particularly significant.